Reverted Plugin Version number back to 1.1.1
Sorry for the change earlier. I understand that it not up to me to modify version. Increasing the version number means a release, and Plugin owner here is the one to decide when to release. So, in this commit I have reverted the version number back to what it was.